Assignment1/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Ollama Web GUI Backend - Production-ready API with security hardening"
requires-python = ">=3.9"
dependencies = [
    "alembic>=1.16.5",
    "bcrypt>=5.0.0",
    "fastapi>=0.120.4",
    "httpx>=0.28.1",
    "pydantic>=2.12.3",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.44",
    "uvicorn>=0.38.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "pytest-mock>=3.14.0",
    "httpx>=0.28.1",  # For test client
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
